+++ packages/react-native-cs/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "react-native-cs",
+  "version": "16.0.0",
+  "private": true,
+  "repository": "facebook/react",
+  "dependencies": {
+    "fbjs": "^0.8.16",
+    "object-assign": "^4.1.1"
+  }
+}
