+++ packages/react-native-rt/src/ReactNativeRTEventEmitter.js
