+++ packages/react-native-rt/src/ReactNativeRTFiberEntry.js
