+++ packages/react-native/index.js
@@ -1,5 +1,4 @@
 'use strict';
 
-// TODO: Move the source into packages
 // TODO: This package is a lie. Maybe rename to react-native-renderer?
-module.exports = require('ReactNativeFiberEntry');
+module.exports = require('./src/ReactNativeFiberEntry');
