+++ packages/react-native/src/NativeMethodsMixin.js
