+++ packages/react-native/src/ReactNativeBridgeEventPlugin.js
