+++ packages/react-native/src/ReactNativeComponent.js
