+++ packages/react-native/src/ReactNativeComponentTree.js
