+++ packages/react-native/src/ReactNativeEventEmitter.js
