+++ packages/react-native/src/ReactNativeFiberEntry.js
