+++ packages/react-native/src/ReactNativeFiberErrorDialog.js
