+++ packages/react-native/src/ReactNativeFiberInspector.js
