+++ packages/react-native/src/ReactNativeFiberRenderer.js
