+++ packages/react-native/src/ReactNativeGlobalResponderHandler.js
