+++ packages/react-native/src/ReactNativeInjection.js
