+++ packages/react-native/src/ReactNativePropRegistry.js
