+++ packages/react-native/src/ReactNativeViewConfigRegistry.js
