+++ packages/react-native/src/__mocks__/BatchedBridge.js
