+++ packages/react-native/src/__mocks__/ExceptionsManager.js
