+++ packages/react-native/src/__mocks__/InitializeCore.js
