+++ packages/react-native/src/__mocks__/Platform.js
