+++ packages/react-native/src/__mocks__/RCTEventEmitter.js
