+++ packages/react-native/src/__mocks__/RTManager.js
