+++ packages/react-native/src/__mocks__/TextInputState.js
