+++ packages/react-native/src/__mocks__/UIManager.js
