+++ packages/react-native/src/__mocks__/View.js
