+++ packages/react-native/src/__mocks__/deepDiffer.js
