+++ packages/react-native/src/__mocks__/deepFreezeAndThrowOnMutationInDev.js
