+++ packages/react-native/src/__mocks__/flattenStyle.js
