+++ packages/react-native/src/__tests__/createReactNativeComponentClass-test.js
