+++ packages/react-native/src/createReactNativeComponentClass.js
