+++ packages/react-native/src/takeSnapshot.js
